server/requirements.txt
uvicorn
fastapi
psycopg2-binary
python_dotenv
pydantic
requests
starlette
bcrypt
python-multipart
